+++ PitchDesktopApp/PitchApp/forms.py
@@ -1,0 +1,7 @@
+from django.forms import ModelForm
+from django.contrib.auth.forms import UserCreationForm
+from .models import User
+
+
+class NormalUserSignUpForm(UserCreationForm):
+    pass
